backend/requirements.txt
# backend/requirements.txt

Flask==2.2.5
gunicorn==20.1.0
python-dotenv==1.0.1
cryptography==41.0.7
SQLAlchemy==2.0.25
psycopg[binary]
minio==7.2.15
kafka-python==1.4.7
Flask-Cors==4.0.0
celery==5.3.6
redis==5.0.1
six==1.16.0
Werkzeug==2.2.3
Flask-JWT-Extended==4.4.0
boto3
pypdf
requests
pgvector
mobi
python-docx
openpyxl
python-pptx
Ebooklib
html2text
pytesseract
Pillow
alembic
gevent
pyclamd==0.4.0 # O la versión más reciente compatible
PyMuPDF
PyMuPDF4LLM
spacy #==3.5.0 # O una versión compatible y más reciente que la 2.3.7
#es_core_news_sm # O el modelo que necesites, asegúrate que sea compatible con spacy 3.x
hvac==1.2.1 # Python client for HashiCorp Vault